sbv-pipeline/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# AI/LLM
openai==1.10.0
anthropic==0.18.0
tiktoken==0.5.2

# Web Scraping & Research
playwright==1.41.0
beautifulsoup4==4.12.3
requests==2.31.0
aiohttp==3.9.1
tavily-python==0.3.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
jsonschema==4.21.1

# Visualization
streamlit==1.30.0
plotly==5.18.0
matplotlib==3.8.2

# Google Sheets
gspread==6.0.0
google-auth==2.27.0
oauth2client==4.1.3

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.6
httpx==0.26.0
tenacity==8.2.3
click==8.1.7

# Jupyter
jupyter==1.0.0
ipykernel==6.29.0
ipywidgets==8.1.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

